% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_offshore.R
\name{clean_offshore}
\alias{clean_offshore}
\title{Title clean offshore bond from Wind}
\usage{
clean_offshore(df)
}
\arguments{
\item{df}{dataframe to be cleaned}
}
\value{
cleaned dataframe
}
\description{
Title clean offshore bond from Wind
}
\examples{
cleaned_offshore(off_jul)
}
