% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_onshore.R
\name{clean_onshore}
\alias{clean_onshore}
\title{Title clean dataframe read from Wind bonds}
\usage{
clean_onshore(df)
}
\arguments{
\item{df}{dataframe to be cleaned}
}
\value{
cleaned dataframe
}
\description{
Title clean dataframe read from Wind bonds
}
\examples{
clean_onshore(allbonds)
}
