% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_show.R
\name{save_pic}
\alias{save_pic}
\title{save ggplot plot with showtext fix}
\usage{
save_pic(filename, height = 11.25)
}
\arguments{
\item{filename}{filename as string}

\item{height}{height of the plot}
}
\value{
Nothing
}
\description{
save ggplot plot with showtext fix
}
\examples{
save_pic("pic.png")
}
