% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_show.R
\name{save_show}
\alias{save_show}
\title{save and show ggplot output from rnotebook.}
\usage{
save_show(pic_name)
}
\arguments{
\item{pic_name}{name of the pic, no quote is needed. if not provided, random name will be used}
}
\value{
nothing.
}
\description{
save and show ggplot output from rnotebook.
}
\examples{
save_show(pic.png)
}
