% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tbl_pic.R
\name{save_tbl_pic}
\alias{save_tbl_pic}
\title{Draw raster plot for the dataframe, used in notebook.}
\usage{
save_tbl_pic(df, file, title, source = "", title_just = 0, note_size = 8)
}
\arguments{
\item{df}{table to draw picture}

\item{file}{file name to save, support quotation, can be missing, random name will be used.}

\item{title}{string, for title}

\item{source}{source of the data to be shown at footnote}

\item{title_just}{a number for just of title, 0 for left, -0.1 for center.}

\item{note_size}{footnote size, adjust this in order to prevent cut off.}
}
\value{
nothing
}
\description{
Draw raster plot for the dataframe, used in notebook.
}
\examples{
save_tbl_pic(mtcars, mtcars.png, title = "mtcars chart", source = "R data")
}
