% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_styled_xlsx.R
\name{write_styled_xlsx}
\alias{write_styled_xlsx}
\title{Saved styled xlsx file}
\usage{
write_styled_xlsx(df, file_name, sheet_name = "sheet1", boldCols = NULL)
}
\arguments{
\item{df}{dataframe to save}

\item{file_name}{file name to save to}

\item{sheet_name}{[optional] name of the sheet to save}

\item{boldCols}{apply bold fonts to text in that column.}
}
\value{
invisble wb object
}
\description{
Saved styled xlsx file
}
\examples{
write_styled_xlsx(mtcars, "mtcars.xlsx", "first", 1:2)

}
